requirements.txt
# Core
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# News
feedparser>=6.0.11

# AI / LLM
anthropic>=0.40.0
openai>=1.50.0
runwayml>=4.0.0

# Video
ffmpeg-python>=0.2.0
Pillow>=10.0.0

# Dashboard
streamlit>=1.30.0

# Scheduling
schedule>=1.2.2

# Speech recognition
openai-whisper>=20240930

# Utilities
aiofiles>=24.1.0
tenacity>=9.0.0

# Social scoring
pytrends>=4.9.2
